jobs/get_audio_video/requirements.txt
google-cloud-storage>=2.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
moviepy>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.0.0
requests>=2.28.0